Criando e usando um contexto global para alteração imediata e sincronização de todas as páginas
