+++ src/contexts/FavoritesContext.tsx
@@ -1,0 +1,108 @@
+import React, { createContext, useContext, useState, useEffect, ReactNode } from 'react';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { Alert } from 'react-native';
+import { Property } from '../components/common/PropertyCard';
+
+const FAVORITES_KEY = '@favorited_properties';
+
+interface FavoritesContextType {
+    favoritedIds: string[];
+    toggleFavorite: (property: Property) => Promise<boolean>;
+    isPropertyFavorite: (propertyId: string) => boolean;
+    clearAllFavorites: () => Promise<void>;
+    loadFavorites: () => Promise<void>;
+}
+
+const FavoritesContext = createContext<FavoritesContextType | undefined>(undefined);
+
+interface FavoritesProviderProps {
+    children: ReactNode;
+}
+
+export const FavoritesProvider: React.FC<FavoritesProviderProps> = ({ children }) => {
+    const [favoritedIds, setFavoritedIds] = useState<string[]>([]);
+
+    // Carregar favoritos ao iniciar
+    useEffect(() => {
+        loadFavorites();
+    }, []);
+
+    const loadFavorites = async () => {
+        try {
+            const savedIds = await AsyncStorage.getItem(FAVORITES_KEY);
+            if (savedIds) {
+                setFavoritedIds(JSON.parse(savedIds));
+            }
+        } catch (error) {
+            console.error('Erro ao carregar favoritos:', error);
+        }
+    };
+
+    const toggleFavorite = async (property: Property) => {
+        let isAlreadyFavorite = favoritedIds.includes(property.id);
+
+        try {
+            let newFavoritedIds: string[];
+
+            if (isAlreadyFavorite) {
+                newFavoritedIds = favoritedIds.filter(id => id !== property.id);
+            } else {
+                newFavoritedIds = [...favoritedIds, property.id];
+            }
+
+            setFavoritedIds(newFavoritedIds);
+            await AsyncStorage.setItem(FAVORITES_KEY, JSON.stringify(newFavoritedIds));
+
+            Alert.alert(
+                isAlreadyFavorite ? 'Removido' : 'Salvo!',
+                isAlreadyFavorite
+                    ? 'Imóvel removido dos salvos'
+                    : 'Imóvel salvo com sucesso!'
+            );
+
+            return !isAlreadyFavorite;
+
+        } catch (error) {
+            console.error('Erro ao salvar imóvel:', error);
+            Alert.alert('Erro', 'Não foi possível salvar o imóvel');
+            return isAlreadyFavorite;
+        }
+    };
+
+    const isPropertyFavorite = (propertyId: string) => {
+        return favoritedIds.includes(propertyId);
+    };
+
+    const clearAllFavorites = async () => {
+        try {
+            await AsyncStorage.removeItem(FAVORITES_KEY);
+            setFavoritedIds([]);
+            Alert.alert('Sucesso', 'Todos os imóveis foram removidos');
+        } catch (error) {
+            console.error('Erro ao limpar favoritos:', error);
+            Alert.alert('Erro', 'Não foi possível limpar os imóveis');
+        }
+    };
+
+    const value: FavoritesContextType = {
+        favoritedIds,
+        toggleFavorite,
+        isPropertyFavorite,
+        clearAllFavorites,
+        loadFavorites
+    };
+
+    return (
+        <FavoritesContext.Provider value={value}>
+            {children}
+        </FavoritesContext.Provider>
+    );
+};
+
+export const useFavoritesContext = () => {
+    const context = useContext(FavoritesContext);
+    if (context === undefined) {
+        throw new Error('useFavoritesContext deve ser usado dentro de um FavoritesProvider');
+    }
+    return context;
+};
