+++ src/screens/App/ExploreScreen.tsx
@@ -15,7 +15,7 @@ export default function ExploreScreen({ navigation }: any) {
     const [maxPrice, setMaxPrice] = useState('');
 
     const { favoritedIds, toggleFavorite, isPropertyFavorite } = useFavorites();
-    
+
     const categories = ['All', 'House', 'Apartment', 'Farm House', 'Shop', 'Villa', 'Condo'];
 
     const filteredProperties = useMemo(() => {
@@ -32,14 +32,14 @@ export default function ExploreScreen({ navigation }: any) {
                 (prop.description && prop.description.toLowerCase().includes(lowercasedSearchText))
             );
         }
-        
+
         if (minPrice !== '') {
             const min = Number(minPrice);
             if (!isNaN(min)) {
                 properties = properties.filter(prop => prop.price >= min);
             }
         }
-        
+
         if (maxPrice !== '') {
             const max = Number(maxPrice);
             if (!isNaN(max)) {
@@ -48,15 +48,15 @@ export default function ExploreScreen({ navigation }: any) {
         }
 
         return properties;
-    }, [selectedCategory, searchText, minPrice, maxPrice, favoritedIds]);
+    }, [selectedCategory, searchText, minPrice, maxPrice]);
 
     const handlePropertyPress = (property: Property) => {
         navigation.navigate('PropertyDetails', { property });
     };
 
     const handleFavoritePress = async (property: Property) => {
-    await toggleFavorite(property);
-};
+        await toggleFavorite(property);
+    };
 
     return (
         <ScreenBackground style={styles.container}>
@@ -92,7 +92,7 @@ export default function ExploreScreen({ navigation }: any) {
                     value={maxPrice}
                 />
             </View>
-            
+
             <View style={styles.categoryContainer}>
                 <FlatList
                     data={categories}
@@ -122,10 +122,8 @@ export default function ExploreScreen({ navigation }: any) {
                     <PropertyCard
                         property={item}
                         onPress={() => handlePropertyPress(item)}
-                        // adicionar a lógica dos favs futuramente aqui <---
                         onFavoritePress={() => handleFavoritePress(item)}
                         isFavorite={isPropertyFavorite(item.id)}
-                        key={item.id + favoritedIds.includes(item.id)}
                     />
                 )}
                 showsVerticalScrollIndicator={false}
